\name{plot.glmm.admb}
\alias{plot.glmm.admb}
\title{Plot Model Fit}
\description{
  Default plotting method for a fitted \code{glmm.admb} model.
  Currently, it only plots the residuals.
}
\usage{
\method{plot}{glmm.admb}(x, \dots)
}
\arguments{
  \item{x}{a fitted \code{glmm.admb} model.}
  \item{\dots}{further arguments passed to \code{plot}.}
}
\seealso{
  \code{\link{glmm.admb}}.

  \code{\link{glmmADMB-package}}.
}
\keyword{models}
